+++ src/command_modules/azure-cli-consumption/azure/cli/command_modules/consumption/_help.py
@@ -21,7 +21,7 @@ helps['consumption reservations summaries'] = """
 """
 helps['consumption reservations summaries list'] = """
     type: command
-    short-summary: List reservations summaries for daily or monthly by reservation order Id and optionally by reservation id.
+    short-summary: List reservations summaries for daily or monthly by order Id or reservation id.
 """
 
 helps['consumption reservations details'] = """
@@ -31,7 +31,7 @@ helps['consumption reservations details'] = """
 
 helps['consumption reservations details list'] = """
     type: command
-    short-summary: List reservations details by reservation order id and optionally by reservation id
+    short-summary: List the details of a reservation by order id or reservation id.
 """
 
 helps['consumption usage'] = """
@@ -46,10 +46,49 @@ helps['consumption usage list'] = """
 
 helps['consumption pricesheet'] = """
     type: group
-    short-summary: Inspect the price sheet by Azure subscription within current or specified billing period.
+    short-summary: Inspect the price sheet of an Azure subscription within a billing period.
 """
 
 helps['consumption pricesheet show'] = """
     type: command
-    short-summary: Show the price sheet for Azure subscription within a current billing period or specified billing period
+    short-summary: Show the price sheet for an Azure subscription within a billing period.
+"""
+
+helps['consumption marketplace'] = """
+    type: group
+    short-summary: Inspect the marketplace usage data of an Azure subscription within a billing period.
+"""
+
+helps['consumption marketplace list'] = """
+    type: command
+    short-summary: Show the marketplace for an Azure subscription within a billing period.
+"""
+
+helps['consumption budget'] = """
+    type: group
+    short-summary: Manage budgets for an Azure subscription.
+"""
+
+helps['consumption budget list'] = """
+    type: command
+    short-summary: List budgets for an Azure subscription.
+"""
+
+helps['consumption budget show'] = """
+    type: command
+    short-summary: Show budget for an Azure subscription.
+"""
+
+helps['consumption budget create'] = """
+    type: command
+    short-summary: Create a budget for an Azure subscription.
+"""
+
+helps['consumption budget update'] = """
+    type: command
+    short-summary: Update a budget for an Azure subscription.
+"""
+helps['consumption budget delete'] = """
+    type: command
+    short-summary: Delete a budget for an Azure subscription.
 """