+++ src/command_modules/azure-cli-consumption/azure/cli/command_modules/consumption/_validators.py
@@ -4,7 +4,17 @@
 # --------------------------------------------------------------------------------------------
 
 from datetime import datetime
+from decimal import Decimal
 from azure.cli.core.util import CLIError
+
+
+def get_decimal_type():
+    def decimal_type(string):
+        try:
+            return Decimal(string)
+        except ValueError:
+            raise ValueError("the value passed cannot be converted to decimal")
+    return decimal_type
 
 
 def get_datetime_type():
@@ -37,3 +47,19 @@ def validate_reservations_summaries(namespace):
         raise CLIError("usage error: --grain  can be either daily or monthly.")
     if data_grain == 'daily' and (not namespace.start_date or not namespace.end_date):
         raise CLIError("usage error: Both --start-date and --end-date need to be supplied for daily grain.")
+
+
+def validate_budget_parameters(namespace):
+    """lowercase the time grain for comparison"""
+
+    budgetcategory = namespace.category.lower()
+    if budgetcategory != 'cost' and budgetcategory != 'usage':
+        raise CLIError("usage error: --parameters.category must be set to either cost or usage")
+
+    time_grain = namespace.time_grain.lower().strip()
+
+    if time_grain != 'annually' and time_grain != 'quarterly' and time_grain != 'monthly':
+        raise CLIError("usage error: --parameters.time_grain can be 'Annually', 'Quarterly', or 'Monthly'.")
+
+    if namespace.amount < 0:
+        raise CLIError("usage error: --parameters.amount must be greater than 0")