+++ src/command_modules/azure-cli-consumption/azure/cli/command_modules/consumption/tests/latest/test_consumption_commands.py
@@ -72,13 +72,34 @@ class AzureConsumptionServiceScenarioTest(ScenarioTest):
         else:
             self.assertIsNone(pricesheet['pricesheets'][0]['meterDetails'])
 
-    @AllowLargeResponse()
+    def _validate_marketplace(self, marketplace, billing_period_id=None):
+        self.assertIsNotNone(marketplace)
+        self.assertTrue(marketplace['id'])
+        self.assertTrue(marketplace['name'])
+        self.assertIsNotNone(marketplace['type'])
+        self.assertIsNotNone(marketplace['instanceName'])
+        self.assertIsNotNone(marketplace['instanceId'])
+        self.assertIsNotNone(marketplace['currency'])
+        self.assertIsNotNone(marketplace['pretaxCost'])
+        self.assertIsNotNone(marketplace['isEstimated'])
+        self.assertIsNotNone(marketplace['orderNumber'])
+        if billing_period_id:
+            self.assertTrue(billing_period_id in marketplace['billingPeriodId'])
+        else:
+            self.assertIsNotNone(marketplace['billingPeriodId'])
+
+    def _validate_budget(self, output_budget):
+        self.assertIsNotNone(output_budget)
+        self.assertTrue(output_budget['amount'])
+        self.assertTrue(output_budget['timeGrain'])
+        self.assertTrue(output_budget['timePeriod'])
+        self.assertTrue(output_budget['name'])
+
     def test_consumption_pricesheet_billing_period(self):
         pricesheet = self.cmd('consumption pricesheet show -p 20171001').get_output_in_json()
         self.assertTrue(pricesheet)
         self._validate_pricesheet(pricesheet, False)
 
-    @AllowLargeResponse()
     def test_consumption_pricesheet(self):
         pricesheet = self.cmd('consumption pricesheet show').get_output_in_json()
         self.assertTrue(pricesheet)
@@ -132,31 +153,69 @@ class AzureConsumptionServiceScenarioTest(ScenarioTest):
         self.assertTrue(all(usage_date_check(usage_date) for usage_date in usages_list))
 
     def test_list_reservations_summaries_monthly(self):
-        reservations_summaries_monthly_list = self.cmd('consumption reservations summaries list -g ''monthly'' -r ca69259e-bd4f-45c3-bf28-3f353f9cce9b').get_output_in_json()
+        reservations_summaries_monthly_list = self.cmd('consumption reservations summaries list -g "monthly" -r ca69259e-bd4f-45c3-bf28-3f353f9cce9b').get_output_in_json()
         self.assertTrue(reservations_summaries_monthly_list)
         self._validate_reservation_summaries(reservations_summaries_monthly_list[0])
 
     def test_list_reservations_summaries_monthly_with_reservationid(self):
-        reservations_summaries_monthly_withreservationid_list = self.cmd('consumption reservations summaries list -g ''monthly'' -r ca69259e-bd4f-45c3-bf28-3f353f9cce9b -i f37f4b70-52ba-4344-a8bd-28abfd21d640').get_output_in_json()
+        reservations_summaries_monthly_withreservationid_list = self.cmd('consumption reservations summaries list -g "monthly" -r ca69259e-bd4f-45c3-bf28-3f353f9cce9b -i f37f4b70-52ba-4344-a8bd-28abfd21d640').get_output_in_json()
         self.assertTrue(reservations_summaries_monthly_withreservationid_list)
         self._validate_reservation_summaries(reservations_summaries_monthly_withreservationid_list[0])
 
     def test_list_reservations_summaries_daily(self):
-        reservations_summaries_daily_list = self.cmd('consumption reservations summaries list -g ''daily'' -r ca69259e-bd4f-45c3-bf28-3f353f9cce9b -s ''2017-12-01'' -e ''2017-12-07''').get_output_in_json()
+        reservations_summaries_daily_list = self.cmd('consumption reservations summaries list -g "daily" -r ca69259e-bd4f-45c3-bf28-3f353f9cce9b -s "2017-12-01" -e "2017-12-07"').get_output_in_json()
         self.assertTrue(reservations_summaries_daily_list)
         self._validate_reservation_summaries(reservations_summaries_daily_list[0])
 
     def test_list_reservations_summaries_daily_with_reservationid(self):
-        reservations_summaries_daily_withreservationid_list = self.cmd('consumption reservations summaries list -g ''daily'' -r ca69259e-bd4f-45c3-bf28-3f353f9cce9b -i f37f4b70-52ba-4344-a8bd-28abfd21d640 -s ''2017-12-01'' -e ''2017-12-07''').get_output_in_json()
+        reservations_summaries_daily_withreservationid_list = self.cmd('consumption reservations summaries list -g "daily" -r ca69259e-bd4f-45c3-bf28-3f353f9cce9b -i f37f4b70-52ba-4344-a8bd-28abfd21d640 -s "2017-12-01" -e "2017-12-07"').get_output_in_json()
         self.assertTrue(reservations_summaries_daily_withreservationid_list)
         self._validate_reservation_summaries(reservations_summaries_daily_withreservationid_list[0])
 
     def test_list_reservations_details(self):
-        reservations_details_list = self.cmd('consumption reservations details list -r ca69259e-bd4f-45c3-bf28-3f353f9cce9b -s ''2017-12-01'' -e ''2017-12-07''').get_output_in_json()
+        reservations_details_list = self.cmd('consumption reservations details list -r ca69259e-bd4f-45c3-bf28-3f353f9cce9b -s "2017-12-01" -e "2017-12-07"').get_output_in_json()
         self.assertTrue(reservations_details_list)
         self._validate_reservation_details(reservations_details_list[0])
 
     def test_list_reservations_details_with_reservationid(self):
-        reservations_details_list = self.cmd('consumption reservations details list -r ca69259e-bd4f-45c3-bf28-3f353f9cce9b -i f37f4b70-52ba-4344-a8bd-28abfd21d640 -s ''2017-12-01'' -e ''2017-12-07''').get_output_in_json()
+        reservations_details_list = self.cmd('consumption reservations details list -r ca69259e-bd4f-45c3-bf28-3f353f9cce9b -i f37f4b70-52ba-4344-a8bd-28abfd21d640 -s "2017-12-01" -e "2017-12-07"').get_output_in_json()
         self.assertTrue(reservations_details_list)
         self._validate_reservation_details(reservations_details_list[0])
+
+    def test_consumption_marketplace_list(self):
+        marketplace_list = self.cmd('consumption marketplace list').get_output_in_json()
+        self.assertTrue(marketplace_list)
+        all(self._validate_marketplace(marketplace_item) for marketplace_item in marketplace_list)
+
+    def test_consumption_marketplace_list_billing_period_filter(self):
+        marketplace_list = self.cmd('consumption marketplace list -p 201804-1 -t 1').get_output_in_json()
+        self.assertTrue(marketplace_list)
+        self.assertTrue(len(marketplace_list) == 1)
+        all(self._validate_marketplace(marketplace_item) for marketplace_item in marketplace_list)
+
+    def test_consumption_marketplace_list_billing_period(self):
+        marketplace_list = self.cmd('consumption marketplace list -p 201804-1').get_output_in_json()
+        self.assertTrue(marketplace_list)
+        all(self._validate_marketplace(marketplace_item, '201804-1') for marketplace_item in marketplace_list)
+
+    def test_consumption_budget_usage_create(self):
+        output_budget = self.cmd('az consumption budget create -b usagetypebudget1 -a 20 -s 2018-02-01 -e 2018-10-01 -t annually -c usage -m 0dfadad2-6e4f-4078-85e1-90c230d4d482').get_output_in_json()
+        self.assertTrue(output_budget)
+        self._validate_budget(output_budget)
+
+    def test_consumption_budget_create(self):
+        output_budget = self.cmd('consumption budget create -b "costbudget" -c "cost" -a 100.0 -s "2018-02-01" -e "2018-10-01" -t "monthly"').get_output_in_json()
+        self.assertTrue(output_budget)
+        self._validate_budget(output_budget)
+
+    def test_consumption_budget_update(self):
+        output_budget = self.cmd('consumption budget create -b "costbudget3" -c "cost" -a 100.0 -s "2018-02-01" -e "2018-10-01" -t "monthly"').get_output_in_json()
+        eTag = output_budget['eTag']
+
+        output = self.cmd('consumption budget update -b "costbudget3" -c "cost" -a 125.0 -s "2018-02-01" -e "2018-10-01" -t "monthly" -x "{}"'.format(eTag)).get_output_in_json()
+        self.assertTrue(output)
+        self.assertEqual(output['amount'], '125')
+
+    def test_consumption_budget_delete(self):
+        output = self.cmd('consumption budget delete -b "costbudget"')
+        self.assertTrue(output)